Added the rest of the render entity files
Added the RenderAnglerFish - RenderWhale files based on how Graeme did
it with the squid man. Not sure if I did it just right because I just
used copy/paste and find/replace to do the dirt work. There will be
errors because the Render files are referring to the Model files for
each mob, which aren't there yet!